bot/requirements.txt
python-telegram-bot>=21.0
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
